backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Acutapugione <acuta.pugione@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.3"}
sqlmodel = "^0.0.22"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.scripts]
start = "main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
